refactor: number 패키지명을 domain으로 변경
Hint DTO가 추가됨에 따라 일관성을 위해 수정
